3.0.1-beta: v2r: resolve #55 "export vocabulary table to CSV".
This is enabled is new drop-down menu next to the class in View mode.
Note, for multi-valued attributes, the new CSV export function uses the `-----` separator in the
corresponding CSV cell. This separator is recognized by the CSV import function to reflect the multiple values.
So, this separator handling is similar as that in the <multivalueedit> directive.